Remove ObjectProxy check that breaks selected for belongsTo items
Removed the check ```!(selected instanceof ObjectProxy)``` from the setter for ```selected```.  This check was added for issue #1147, but completely breaks powerselect's initial selection when dealing with an object that is 'async'.

The primary case here is when we pull an object out of a 'belongsTo' relationship, and use that as the 'selected' value to pass to powerselect; powerselect DOES have the code necessary to resolve the promise and deal with stuff, but because Ember Data passes it as an ObjectProxy, it completely skips this code!

Since we already have the check for "does it have a .then", why do we need the shortcircuiting on instanceof ObjectProxy?